Adding new custom-templates, and registry identity for ITE. (Updated for SonarCloud analysis)
