+++ src/components/CertificateTemplates/tlds/sg/edu/ite/common/certificate.js
@@ -5,7 +5,6 @@ import {
   IMG_CERT_NIEC1_LOGO_ITE,
   IMG_CERT_NIEC1ITE_LOGO_ITE
 } from "./images";
-import { formatDate, formatDatePrefix, formatCertID } from "./functions";
 
 export const fullWidthStyle = {
   width: "100%",
@@ -194,6 +193,28 @@ export const renderTwoSignatures = certificate => (
   </div>
 );
 
+export const renderITEFooter = certificate => (
+  <div className="container">
+    <div
+      className="row d-flex justify-content-start align-items-start"
+      style={{ marginTop: "1rem" }}
+    >
+      <div className="col-1" />
+      <div className="col-10 text-left">
+        <p style={arial5Pt}>{certificate.additionalData.footnote1}</p>
+        <div className="ml-3">
+          <div className="pl-4">
+            <p style={arial5Pt}>{certificate.additionalData.footnote2}</p>
+          </div>
+        </div>
+        <br />
+        <br />
+      </div>
+      <div className="col-1" />
+    </div>
+  </div>
+);
+
 export const renderTwoNiecSignatures = certificate => (
   <div
     className="row d-flex justify-content-center"
@@ -284,7 +305,7 @@ export const renderTwoNiecSignatures = certificate => (
   </div>
 );
 
-const renderAwardText = certificate => (
+export const renderFullCertAwardText = certificate => (
   <div>
     <div
       className="row d-flex justify-content-center"
@@ -376,29 +397,96 @@ const renderAwardText = certificate => (
   </div>
 );
 
-export const renderIssuingDate = certificate => (
+export const renderCOMAwardText = certificate => (
   <div>
-    <p>
-      {formatDatePrefix(certificate.issuedOn)}{" "}
-      {formatDate(certificate.issuedOn)}
-    </p>
-  </div>
-);
-
-export const renderFooter = certificate => (
-  <div className="container">
+    <div
+      className="row d-flex justify-content-center"
+      style={{ marginTop: "2rem" }}
+    >
+      <span style={arial16Pt}>It is hereby certified that</span>
+    </div>
+    <div className="ml-3">
+      <div className="mr-3">
+        <div className="row d-flex justify-content-center">
+          <span style={timesNewRoman32Pt}>{certificate.recipient.name}</span>
+        </div>
+      </div>{" "}
+    </div>
     <div className="row d-flex justify-content-center">
-      <div className="col-6 text-left">
-        {get(certificate, "additionalData.additionalCertId")}
-      </div>
-      <div className="col-6 text-right">{formatCertID(certificate.id)}</div>
+      <span style={arial10Pt}>
+        __________________________________________________________________________________________
+      </span>
+    </div>
+    <p>
+      <br />
+    </p>
+    <div className="row d-flex justify-content-center">
+      <span style={arial16Pt}>was awarded the Certificate of Merit</span>
+    </div>
+    <div className="row d-flex justify-content-center">
+      <span style={arial16Pt}>for Outstanding Performance in the</span>
+    </div>
+    <p>
+      <br />
+    </p>
+    <div className="ml-5">
+      <div className="mr-5">
+        <div className="row d-flex justify-content-center">
+          <span style={timesNewRoman32Pt}>
+            <p style={timesNewRoman32Pt}>
+              {certificate.additionalData.certDescr1}
+            </p>
+          </span>
+        </div>
+      </div>{" "}
+    </div>
+    <div className="ml-5">
+      <div className="mr-5">
+        <div className="row d-flex justify-content-center">
+          <span style={timesNewRoman32Pt}>
+            <p style={timesNewRoman32Pt}>
+              {certificate.additionalData.certDescr2}
+            </p>
+          </span>
+        </div>
+      </div>{" "}
+    </div>
+    <div className="ml-5">
+      <div className="mr-5">
+        <div className="row d-flex justify-content-center">
+          <span style={timesNewRoman32Pt}>
+            <p style={timesNewRoman32Pt}>
+              {certificate.additionalData.certDescr3}
+            </p>
+          </span>
+        </div>
+      </div>{" "}
+    </div>
+    <div className="ml-5">
+      <div className="mr-5">
+        <div className="row d-flex justify-content-center">
+          <span style={timesNewRoman32Pt}>
+            <p style={timesNewRoman32Pt}>
+              {certificate.additionalData.certDescr4}
+            </p>
+          </span>
+        </div>
+      </div>{" "}
+    </div>
+    <div className="row d-flex justify-content-center">
+      <span style={arial16Pt}>on</span>
+    </div>
+    <div className="row d-flex justify-content-center">
+      <span style={timesNewRoman32Pt}>
+        <p style={timesNewRoman32Pt}>{certificate.additionalData.confDate}</p>
+      </span>
     </div>
   </div>
 );
 
 /* eslint-disable */
 // Disabled eslint as there's no way to add proptypes to an anonymous function like this
-export default ({}) => ({ certificate }) => (
+export default () => ({ certificate }) => (
   <div>
     <div
       className="container"
@@ -408,9 +496,10 @@ export default ({}) => ({ certificate }) => (
       {renderAwardText(certificate)}
       {certificate.additionalData.certSignatories &&
       certificate.additionalData.certSignatories[2]
-        ? renderThreeSignatures(certificate)
-        : renderTwoSignatures(certificate)}
+        ? renderTwoSignatures(certificate)
+        : renderTwoNiecSignatures(certificate)}
 	  {renderITEFooter(certificate)}
-    </div>
+
+  </div>
   </div>
 );
